1833 add find user by id

+++ SkillRocks_test/src/main/java/com/example/SkillRocks_test/service/UserServiceImpl.java
@@ -50,9 +50,12 @@ public class UserServiceImpl implements UserService {
     @Transactional(propagation = Propagation.REQUIRED)
     public void delete (UUID id) {
         findById(id);
-        userRepository.deleteById(id);
+        if (checkId(id)){
+            userRepository.deleteById(id);
+        }
     }
 
-
-
+    private boolean checkId(UUID id) {
+        return id != null;
+    }
 }
